--- src/main/java/com/akatsuki/newsum/common/dto/pagination/model/cursor/Cursor.java
@@ -1,5 +1,0 @@
-package com.akatsuki.newsum.common.dto.pagination.model.cursor;
-
-public interface Cursor {
-	String getCursor();
-}
