+++ src/main/java/com/akatsuki/newsum/common/pagination/model/cursor/Cursor.java
@@ -1,0 +1,5 @@
+package com.akatsuki.newsum.common.pagination.model.cursor;
+
+public interface Cursor {
+	String getCursor();
+}
