+++ src/main/java/com/akatsuki/newsum/common/pagination/model/page/CursorPage.java
@@ -1,4 +1,4 @@
-package com.akatsuki.newsum.common.dto.pagination.model;
+package com.akatsuki.newsum.common.pagination.model.page;
 
 import lombok.Getter;
